prepare reflex setup
