+++ .cruft.json
@@ -4,7 +4,7 @@
   "context": {
     "cookiecutter": {
       "project_name": "drop",
-      "short_summary": "Data upload and download API for the MEx project.",
+      "short_summary": "Data upload and download service for the MEx project.",
       "long_summary": "The `mex-drop` package provides an API for uploading data to and downloading data from the MEx project. Request payloads need to be JSON-formatted but can have arbitrary structures. Accepted data will be ingested and processed asynchronously.",
       "_template": "https://github.com/robert-koch-institut/mex-template"
     }
