+++ Dockerfile
@@ -3,7 +3,7 @@
 FROM python:3.11 AS base
 
 LABEL org.opencontainers.image.authors="mex@rki.de"
-LABEL org.opencontainers.image.description="Data upload and download API for the MEx project."
+LABEL org.opencontainers.image.description="Data upload and download service for the MEx project."
 LABEL org.opencontainers.image.licenses="MIT"
 LABEL org.opencontainers.image.url="https://github.com/robert-koch-institut/mex-drop"
 LABEL org.opencontainers.image.vendor="robert-koch-institut"
@@ -15,7 +15,13 @@ ENV PIP_PROGRESS_BAR=off
 ENV PIP_PREFER_BINARY=on
 ENV PIP_DISABLE_PIP_VERSION_CHECK=on
 
-ENV MEX_DROP_HOST=0.0.0.0
+ENV APP_NAME=mex
+ENV FRONTEND_PORT=8020
+ENV DEPLOY_URL=http://0.0.0.0:8020
+ENV BACKEND_PORT=8021
+ENV API_URL=http://0.0.0.0:8021
+ENV TELEMETRY_ENABLED=False
+ENV REFLEX_ENV_MODE=prod
 
 WORKDIR /app
 
@@ -33,6 +39,7 @@ RUN --mount=type=cache,target=/root/.cache/pip pip install .
 
 USER mex
 
-EXPOSE 8081
+EXPOSE 8020
+EXPOSE 8021
 
-ENTRYPOINT [ "drop" ]
+ENTRYPOINT [ "drop", "run" ]
