+++ README.md
@@ -1,6 +1,6 @@
 # MEx drop
 
-Data upload and download API for the MEx project.
+Data upload and download service for the MEx project.
 
 [![cookiecutter](https://github.com/robert-koch-institut/mex-drop/actions/workflows/cookiecutter.yml/badge.svg)](https://github.com/robert-koch-institut/mex-template)
 [![cve-scan](https://github.com/robert-koch-institut/mex-drop/actions/workflows/cve-scan.yml/badge.svg)](https://github.com/robert-koch-institut/mex-drop/actions/workflows/cve-scan.yml)
@@ -103,4 +103,4 @@ components of the MEx project are open-sourced under the same license as well.
 
 ### drop
 
-- `drop` starts the drop service
+- `pdm run drop run` starts the drop service
