+++ pyproject.toml
@@ -1,31 +1,30 @@
 [project]
 name = "mex-drop"
 version = "0.5.0"
-description = "Data upload and download API for the MEx project."
+description = "Data upload and download service for the MEx project."
 authors = [{ name = "MEx Team", email = "mex@rki.de" }]
 readme = { file = "README.md", content-type = "text/markdown" }
 license = { file = "LICENSE" }
 urls = { Repository = "https://github.com/robert-koch-institut/mex-drop" }
-requires-python = "<3.13,>=3.11"
+requires-python = "<3.12,>=3.11"
 dependencies = [
-    "fastapi[standard]==0.112.0",
-    "jinja2==3.1.4",
     "mex-common @ git+https://github.com/robert-koch-institut/mex-common.git@0.33.0",
-    "uvicorn[standard]==0.30.5",
+    "reflex==0.5.8",
+    "uvicorn==0.20.0",
 ]
 optional-dependencies.dev = [
     "ipdb==0.13.13",
     "mypy==1.11.1",
     "pytest-cov==5.0.0",
+    "pytest-playwright==0.5.1",
     "pytest-random-order==1.1.1",
-    "pytest-xdist==3.6.1",
     "pytest==8.3.2",
     "ruff==0.5.6",
     "sphinx==8.0.2",
 ]
 
 [project.scripts]
-drop = "mex.drop.main:main"
+drop = "reflex.reflex:cli"
 
 [tool.cruft]
 template = "https://github.com/robert-koch-institut/mex-template"
@@ -40,13 +39,19 @@ show_error_codes = true
 strict = true
 plugins = ["pydantic.mypy"]
 
+[[tool.mypy.overrides]]
+module = "reflex.*"
+ignore_missing_imports = true
+
 [tool.pdm]
 distribution = true
 
 [tool.pdm.scripts]
 update-all = { cmd = "pdm update --group :all --update-all --save-compatible" }
 lock-all = { cmd = "pdm lock --group :all --python='==3.11.*'" }
-install-all = { cmd = "pdm install --group :all --frozen-lockfile" }
+install-lockfile = { cmd = "pdm install --group :all --frozen-lockfile" }
+install-playwright = { cmd = "playwright install firefox" }
+install-all = { composite = ["install-lockfile", "install-playwright"] }
 export-all = { cmd = "pdm export --group :all --no-hashes -f requirements" }
 apidoc = { cmd = "pdm run sphinx-apidoc -f -o docs/source mex" }
 sphinx = { cmd = "pdm run sphinx-build -aE -b dirhtml docs docs/dist" }
@@ -55,7 +60,7 @@ wheel = { cmd = "pdm build --no-sdist" }
 mypy-daemon = { cmd = "pdm run dmypy run --timeout 7200 -- mex" }
 lint = { cmd = "pre-commit run --all-files" }
 unit = { cmd = "pdm run pytest -m 'not integration'" }
-test = { cmd = "pdm run pytest --numprocesses=auto --dist=worksteal" }
+test = { cmd = "pdm run pytest" }
 all = { composite = ["install-all", "lint", "test", "doc"] }
 
 [tool.pydantic-mypy]
@@ -69,6 +74,7 @@ addopts = [
     "--cov-report=term-missing:skip-covered",
     "--cov-fail-under=95",
     "--cov-branch",
+    "--browser=firefox",
     "--pdbcls=IPython.terminal.debugger:TerminalPdb",
     "--random-order-bucket=global",
 ]
