+++ requirements.txt
@@ -3,3 +3,5 @@ mex-release @ git+https://github.com/robert-koch-institut/mex-release.git
 pdm==2.17.3
 pre-commit==3.8.0
 wheel==0.44.0
+
+reflex==0.5.8
