+++ rxconfig.py
@@ -1,0 +1,3 @@
+import reflex as rx
+
+config = rx.Config(app_name="mex", telemetry_enabled=False)
